connect: carve ui.rs (pure move)
The served surface leaves main.rs: landing/connect/trust/access pages
and their HTML builders, embedded installers and brand assets, health
probes, and the static-asset fallback, with their tests. Items widen to
pub(crate); main.rs mounts the slice as a sibling module with a glob
re-export so every existing reference keeps compiling.
